Update functions-bindings-event-iot.md
The current documentation for IoT Hub Trigger is almost a copy of the Event Hub trigger. Although, IoT Hub API is based on Event Hub API, there are certain values for IoT Hub that are different or obtained in a different location.

After creating IoT Hub, users should be clear about what values to use for EventHubName and Connection, which are different than the ones in Event Hub

For example, the attribute EventHubName for a IoT Hub is named differently and located in Azure Portal of a IoT Hub under the Built-in endpoints section. There is not a reference of the name Event Hub-compatible name in this article. 
It's the same regarding the connection string which in the article is pointing to Event Hub connection string and not the IoT Hub one.

I added above an example for Connections section  and I ask you to kindly review it. I also think the Attributes section should be edited as well and instead of using EventHubName, it should be Event Hub-compatible name instead and then point to the article: https://learn.microsoft.com/en-us/azure/iot-hub/iot-hub-devguide-messages-read-builtin